# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}": {
      "get": {
        "operationId": "WorkspacesController_get_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspacesController_update_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "delete": {
        "operationId": "WorkspacesController_remove_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/projects/{projectId}/tasks": {
      "post": {
        "operationId": "TasksController_create_v1",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks": {
      "get": {
        "operationId": "TasksController_list_v1",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusCategory",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "TODO",
                "IN_PROGRESS",
                "DONE"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "STORY",
                "BUG",
                "TASK",
                "EPIC",
                "SUBTASK"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "URGENT",
                "HIGH",
                "MEDIUM",
                "LOW",
                "NONE"
              ]
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "description": "`me`, `none` (unassigned) or a user id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reporter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "`none` for top-level tasks, or a parent task id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Matches the title, or the key / number (`SYN-12`, `12`).",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "dueBefore",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueAfter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "excludeSubtasks",
            "required": false,
            "in": "query",
            "description": "Hide sub-tasks (boards do this by default).",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "description": "Apply a saved view; explicit parameters override its filters.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "number",
                "createdAt",
                "updatedAt",
                "position",
                "title",
                "priority",
                "startDate",
                "dueDate"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lists tasks across the projects the caller can see; filter with `projectId`, `assignee=me`, etc.",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/projects/{projectId}/board": {
      "get": {
        "operationId": "TasksController_board_v1",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "STORY",
                "BUG",
                "TASK",
                "EPIC",
                "SUBTASK"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "URGENT",
                "HIGH",
                "MEDIUM",
                "LOW",
                "NONE"
              ]
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "description": "`me`, `none` (unassigned) or a user id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reporter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parent",
            "required": false,
            "in": "query",
            "description": "`none` for top-level tasks, or a parent task id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Matches the title, or the key / number (`SYN-12`, `12`).",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "dueBefore",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dueAfter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "excludeSubtasks",
            "required": false,
            "in": "query",
            "description": "Hide sub-tasks (boards do this by default).",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "description": "Apply a saved view; explicit parameters override its filters.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "number"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kanban board of a project: columns in workflow order, cards in rank order.",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/recent-tasks": {
      "get": {
        "operationId": "TasksController_recentTasks_v1",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tasks the caller opened recently, newest first.",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/bulk": {
      "patch": {
        "operationId": "TasksController_bulk_v1",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateTasksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}": {
      "get": {
        "operationId": "TasksController_get_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/archive": {
      "post": {
        "operationId": "TasksController_archive_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/restore": {
      "post": {
        "operationId": "TasksController_restore_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/rank": {
      "post": {
        "operationId": "TasksController_rank_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Drop a card into a column at a position (above/below another card, or at the bottom).",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/viewed": {
      "post": {
        "operationId": "TasksController_viewed_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/duplicate": {
      "post": {
        "operationId": "TasksController_duplicate_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/move": {
      "post": {
        "operationId": "TasksController_move_v1",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Synqonix API",
    "description": "Project management for developers",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WorkspaceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "role": {
            "description": "The caller's role in this workspace.",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER",
              "VIEWER"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "createdAt",
          "role"
        ]
      },
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "type": {
            "type": "string",
            "enum": [
              "STORY",
              "BUG",
              "TASK",
              "EPIC",
              "SUBTASK"
            ]
          },
          "description": {
            "type": "string",
            "description": "Markdown, up to 50,000 characters.",
            "maxLength": 50000
          },
          "statusId": {
            "type": "string",
            "description": "Defaults to the project's first \"to do\" status."
          },
          "priority": {
            "type": "string",
            "enum": [
              "URGENT",
              "HIGH",
              "MEDIUM",
              "LOW",
              "NONE"
            ]
          },
          "assigneeIds": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "labelIds": {
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimate": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "startDate": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": true,
            "description": "Map of custom field id to value."
          }
        },
        "required": [
          "title"
        ]
      },
...
      "TaskListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
...
      "BoardDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnDto"
            }
          },
          "epics": {
            "description": "Epics of the project, for epic swimlanes.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRefDto"
            }
          }
        },
        "required": [
          "projectId",
          "columns",
          "epics"
        ]
      },
...
      "BulkUpdateTasksDto": {
        "type": "object",
        "properties": {
          "taskIds": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "changes": {
            "$ref": "#/components/schemas/BulkChangesDto"
          }
        },
        "required": [
          "taskIds",
          "changes"
        ]
      },
      "BulkResultDto": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "number"
          }
        },
        "required": [
          "updated"
        ]
      },
...
      "RankTaskDto": {
        "type": "object",
        "properties": {
          "statusId": {
            "type": "string",
            "description": "Column to drop the card into. Defaults to the card's current status."
          },
          "beforeId": {
            "type": "string",
            "description": "Place the card directly above this card (which must be in the target column)."
          },
          "afterId": {
            "type": "string",
            "description": "Place the card directly below this card (which must be in the target column)."
          }
        }
      },
      "MoveTaskDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Destination project in the same workspace."
          }
        },
        "required": [
          "projectId"
        ]
      },
...
    }
  }
}